merge main into add-copies

+++ react_library_backend/controllers/users.js
@@ -2,6 +2,7 @@ const bcrypt = require("bcrypt");
 const { request, response } = require("express");
 const userRouter = require("express").Router();
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 userRouter.post("/", async (request, response) => {
   const { email, password, admin } = request.body;
@@ -32,7 +33,12 @@ userRouter.post("/", async (request, response) => {
 });
 
 userRouter.get("/", async (request, response) => {
-  console.log("asd");
+  const authorization = request.get("authorization");
+  const user = jwt.verify(authorization, process.env.SECRET);
+  if (!user.admin) {
+    return response.status(401).json({ error: "no admin rights" });
+  }
+
   const users = await User.find({}).populate("loaned", {
     title: 1,
     author: 1,
@@ -43,7 +49,13 @@ userRouter.get("/", async (request, response) => {
 
 // Get user by ID
 userRouter.get("/:id", async (request, response) => {
-  const user = await User.findById(request.params.id);
+  const authorization = request.get("authorization");
+  let user = jwt.verify(authorization, process.env.SECRET);
+  console.log(user, request.params.id);
+  if (!user.admin && request.params.id !== user.id) {
+    return response.status(401).json({ error: "no rights" });
+  }
+  user = await User.findById(request.params.id);
   if (user) {
     response.json(user.toJSON());
   } else {
@@ -52,20 +64,48 @@ userRouter.get("/:id", async (request, response) => {
 });
 
 userRouter.delete("/:id", async (request, response) => {
+  const authorization = request.get("authorization");
+  let user = jwt.verify(authorization, process.env.SECRET);
+  console.log(user, request.params.id);
+  if (!user.admin && request.params.id !== user.id) {
+    return response.status(401).json({ error: "no rights" });
+  }
+
   await User.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
 
 userRouter.put("/:id", (request, response) => {
-  const body = request.body;
+  const authorization = request.get("authorization");
+  let user = jwt.verify(authorization, process.env.SECRET);
+  console.log(user, request.params.id);
+  if (!user.admin && request.params.id !== user.id) {
+    return response.status(401).json({ error: "no rights" });
+  }
 
-  //   const user = {
-  //     email: body.email,
-  //     loaned: body.loaned,
-  //   };
+  const body = request.body;
 
   User.findByIdAndUpdate(request.params.id, { email: body.email }).then(
     (updatedUser) => {
+      response.json(updatedUser);
+    }
+  );
+});
+
+userRouter.put("/admin/:id", (request, response) => {
+  const authorization = request.get("authorization");
+  let user = jwt.verify(authorization, process.env.SECRET);
+  console.log(user, request.params.id);
+  if (!user.admin) {
+    return response.status(401).json({ error: "no admin rights" });
+  }
+
+  const body = request.body;
+  console.log(body);
+
+  User.findByIdAndUpdate(request.params.id, { admin: body.admin }).then(
+    (updatedUser) => {
+      console.log(updatedUser);
       response.json(updatedUser);
     }
   );
